* 發布第2.2.1版；         * 刪隊gwm/man/zh_HK/man1/Makefile中的調試語句；         * 爲了更好地支持zh_TW和en_US，修改man和po的相關Makefile；         * 絕大部分尺寸改爲用int類型，以免有符號數與無符號數混算；         * 微調文字與構件的間距；         * 構件高度改爲自動由字體尺寸確定。

+++ src/misc.h
@@ -31,7 +31,7 @@ KeySym look_up_key(XIC xic, XKeyEvent *e, wchar_t *keyname, size_t n);
 char *copy_string(const char *s);
 char *copy_strings(const char *s, ...);
 void vfree(void *ptr, ...);
-File *get_files_in_paths(const char *paths, const char *regex, Order order, bool is_fullname, size_t *n);
+File *get_files_in_paths(const char *paths, const char *regex, Order order, bool is_fullname, int *n);
 void free_files(File *head);
 int base_n_floor(int x, int n);
 int base_n_ceil(int x, int n);
